Enabled the json data to be initialised into crypto models when seeds file is run. Crypto data is now available - just need to set it up so it scrapes periodically

+++ db/migrate/20210412115808_create_cryptos.rb
@@ -3,11 +3,11 @@ class CreateCryptos < ActiveRecord::Migration[6.1]
     create_table :cryptos do |t|
       t.integer :cmc_rank
       t.string :name
-      t.float :price
-      t.integer :volume_24h
-      t.integer :market_cap
-      t.integer :circulating_supply
-      t.integer :max_supply
+      t.integer :price
+      t.bigint :volume_24h
+      t.bigint :market_cap
+      t.bigint :circulating_supply
+      t.bigint :max_supply
       
       t.timestamps
     end
